refactor(discovery): delete the pipeline orchestrator (P7 retirement R2)
Completes the discovery convergence (ADR-0007): with the UI on the engine
+ jobs spine and the /security/pipeline/* endpoints gone (R1), the second
orchestrator is fully removed. One orchestrator remains — the engine.

Deleted (~6k lines): pipeline.go, pipeline_phases.go, pipeline_types.go,
the four phase_*.go runners (enumeration/resolution/scanning/assessment)
+ their tests, and the pipeline test suites. Excised the Service's
pipeline coordination layer (SetPipeline/SetOnPipelineComplete/StartPipeline/
GetPipelineStatus/IsPipelineRunning/syncPipelineResults + the pipeline/
onPipelineComplete fields + the PipelineStatus/PipelinePhase status fields
and the broadcast.go lines that read them). Removed the frontend
useCardState "pipeline" event branch + its PipelineEvent types (the backend
no longer emits them).

Relocations (engine still depends on these): pipeline_defaults.go ->
scan_config.go via git mv (preserves the canonical port lists), keeping
GetQuickPorts/GetStandardPorts/GetComprehensivePorts and adding the
PortScanIntensity/ScanTimingProfile enums + SNMPMIBSelection moved out of
the deleted pipeline_types.go; severityCritical relocated into cve_kev.go.
scan_config_test.go ports the enum/port-list coverage; the rest of the
pipeline tests covered now-deleted types.

Gates: go build ./... + CGO_ENABLED=0, golangci 0, go test -race
./internal/discovery/ (incl. scan_config_test), golden HTTP harness
(unchanged — no routes here), ui tsc + biome + 185 vitest, token gate.

Refs ADR-0007
