--- internal/discovery/phase_enumeration.go
@@ -1,684 +1,0 @@
-package discovery
-
-// This file implements Phase 1 (Enumeration) of the discovery pipeline.
-// It wraps existing scanners to provide comprehensive device enumeration.
-
-import (
-	"context"
-	"fmt"
-	"net"
-	"sync"
-	"time"
-
-	"github.com/krisarmstrong/seed/internal/logging"
-)
-
-// Timing constants for enumeration phase.
-const (
-	enumProgressTickerMs   = 500 // Progress reporting interval in milliseconds
-	enumSlowScanDelayS     = 2   // Delay between slow scan passes in seconds
-	enumRetryDelayMs       = 500 // Delay between retry attempts in milliseconds
-	enumSubnetExcludeNum   = 2   // Number of addresses excluded from subnet (network + broadcast)
-	enumDefaultScanTimeMin = 5   // Default scan timeout in minutes
-	enumDefaultPingWorkers = 50  // Default number of concurrent ping workers
-	enumSubnet24HostCount  = 254 // Usable hosts in a /24 subnet for progress estimation
-	enumPercentComplete    = 100 // Value representing 100% completion
-)
-
-// EnumerationPhase implements the Phase interface for device enumeration.
-// Phase 1 discovers ALL devices on local and target subnets using:
-//   - ARP scan (local subnet - Layer 2)
-//   - ICMP ping sweep (additional subnets - Layer 3)
-//   - NDP discovery (IPv6)
-//   - Passive LLDP/CDP/EDP data merge
-type EnumerationPhase struct {
-	deviceDiscovery *DeviceDiscovery
-	config          *EnumerationConfig
-	broadcaster     EventBroadcaster
-}
-
-// EnumerationConfig controls Phase 1 behavior.
-type EnumerationConfig struct {
-	// ARPScan enables ARP scanning for local subnet discovery.
-	ARPScan bool `yaml:"arp_scan" json:"arpScan"`
-
-	// ICMPScan enables ICMP ping sweep for remote subnets.
-	ICMPScan bool `yaml:"icmp_scan" json:"icmpScan"`
-
-	// NDPScan enables IPv6 Neighbor Discovery Protocol scanning.
-	NDPScan bool `yaml:"ndp_scan" json:"ndpScan"`
-
-	// PassiveProtocols controls which passive discovery protocols are active.
-	PassiveProtocols struct {
-		LLDP bool `yaml:"lldp" json:"lldp"`
-		CDP  bool `yaml:"cdp" json:"cdp"`
-		EDP  bool `yaml:"edp" json:"edp"`
-	} `yaml:"passive_protocols" json:"passiveProtocols"`
-
-	// Timing controls scan rate limiting.
-	Timing EnumerationTiming `yaml:"timing" json:"timing"`
-
-	// AdditionalSubnets to scan beyond local subnet.
-	AdditionalSubnets []string `yaml:"additional_subnets" json:"additionalSubnets"`
-}
-
-// EnumerationTiming controls scan delays and concurrency.
-type EnumerationTiming struct {
-	// PingTimeout for individual ICMP requests.
-	PingTimeout time.Duration `yaml:"ping_timeout" json:"pingTimeout"`
-
-	// ScanTimeout for the entire enumeration phase.
-	ScanTimeout time.Duration `yaml:"scan_timeout" json:"scanTimeout"`
-
-	// PingWorkers for concurrent ICMP pinging.
-	PingWorkers int `yaml:"ping_workers" json:"pingWorkers"`
-
-	// ARPDelay between ARP probes (rate limiting).
-	ARPDelay time.Duration `yaml:"arp_delay" json:"arpDelay"`
-}
-
-// DefaultEnumerationConfig returns default enumeration settings.
-func DefaultEnumerationConfig() *EnumerationConfig {
-	return &EnumerationConfig{
-		ARPScan:  true,
-		ICMPScan: true,
-		NDPScan:  true,
-		PassiveProtocols: struct {
-			LLDP bool `yaml:"lldp" json:"lldp"`
-			CDP  bool `yaml:"cdp" json:"cdp"`
-			EDP  bool `yaml:"edp" json:"edp"`
-		}{
-			LLDP: true,
-			CDP:  true,
-			EDP:  true,
-		},
-		Timing: EnumerationTiming{
-			PingTimeout: 1 * time.Second,
-			ScanTimeout: enumDefaultScanTimeMin * time.Minute,
-			PingWorkers: enumDefaultPingWorkers,
-			ARPDelay:    0, // No delay by default (ARP is very fast)
-		},
-	}
-}
-
-// NewEnumerationPhase creates a new Phase 1 implementation.
-func NewEnumerationPhase(
-	deviceDiscovery *DeviceDiscovery,
-	config *EnumerationConfig,
-	broadcaster EventBroadcaster,
-) *EnumerationPhase {
-	if config == nil {
-		config = DefaultEnumerationConfig()
-	}
-	return &EnumerationPhase{
-		deviceDiscovery: deviceDiscovery,
-		config:          config,
-		broadcaster:     broadcaster,
-	}
-}
-
-// Name returns the phase name.
-func (p *EnumerationPhase) Name() string {
-	return "enumeration"
-}
-
-// Run executes the enumeration phase.
-// This is the primary device discovery phase that finds all devices.
-func (p *EnumerationPhase) Run(
-	ctx context.Context,
-	_ []*DiscoveredDevice,
-	progressCh chan<- PhaseProgressPayload,
-) ([]*DiscoveredDevice, error) {
-	start := time.Now()
-	logging.GetLogger().InfoContext(ctx, "Enumeration phase starting",
-		"arp", p.config.ARPScan,
-		"icmp", p.config.ICMPScan,
-		"ndp", p.config.NDPScan)
-
-	scanCtx := p.createScanContext(ctx)
-
-	var progress EnumerationProgress
-	progress.Start()
-
-	done := p.startProgressReporter(progressCh, &progress, start)
-	p.runParallelScans(ctx, scanCtx, &progress)
-	close(done)
-
-	devices := p.deviceDiscovery.GetDevices()
-	p.broadcastDiscoveries(devices)
-
-	logging.GetLogger().InfoContext(ctx, "Enumeration phase completed",
-		"devices", len(devices),
-		"duration", time.Since(start))
-
-	return devices, nil
-}
-
-// createScanContext creates a context with timeout if configured.
-func (p *EnumerationPhase) createScanContext(ctx context.Context) context.Context {
-	if p.config.Timing.ScanTimeout > 0 {
-		scanCtx, cancel := context.WithTimeout(ctx, p.config.Timing.ScanTimeout)
-		// Note: caller should handle cancel, but for enumeration phase the context
-		// is only used during the scan and we rely on parent context cancellation
-		go func() {
-			<-ctx.Done()
-			cancel()
-		}()
-		return scanCtx
-	}
-	return ctx
-}
-
-// startProgressReporter starts a goroutine that reports progress periodically.
-// Returns a channel that should be closed when scanning is complete.
-func (p *EnumerationPhase) startProgressReporter(
-	progressCh chan<- PhaseProgressPayload,
-	progress *EnumerationProgress,
-	start time.Time,
-) chan struct{} {
-	done := make(chan struct{})
-	go func() {
-		ticker := time.NewTicker(enumProgressTickerMs * time.Millisecond)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-done:
-				return
-			case <-ticker.C:
-				p.sendProgressUpdate(progressCh, progress, start)
-			}
-		}
-	}()
-	return done
-}
-
-// sendProgressUpdate sends a progress update if the channel is available.
-func (p *EnumerationPhase) sendProgressUpdate(
-	progressCh chan<- PhaseProgressPayload,
-	progress *EnumerationProgress,
-	start time.Time,
-) {
-	if progressCh != nil {
-		progressCh <- PhaseProgressPayload{
-			Phase:           "enumeration",
-			ProcessedCount:  progress.DevicesFound(),
-			TotalCount:      progress.EstimatedTotal(),
-			PercentComplete: progress.PercentComplete(),
-			CurrentTarget:   progress.CurrentTarget(),
-			ElapsedMs:       time.Since(start).Milliseconds(),
-		}
-	}
-}
-
-// runParallelScans executes all configured scans in parallel.
-func (p *EnumerationPhase) runParallelScans(
-	ctx context.Context,
-	scanCtx context.Context,
-	progress *EnumerationProgress,
-) {
-	var wg sync.WaitGroup
-
-	if p.config.ARPScan {
-		wg.Go(func() {
-			p.executeARPScan(ctx, scanCtx, progress)
-		})
-	}
-
-	if p.config.ICMPScan && len(p.config.AdditionalSubnets) > 0 {
-		wg.Go(func() {
-			p.executeICMPScan(ctx, scanCtx, progress)
-		})
-	}
-
-	if p.config.NDPScan {
-		wg.Go(func() {
-			progress.SetPhase("ndp_scan")
-			progress.MarkComplete("ndp_scan")
-		})
-	}
-
-	wg.Wait()
-}
-
-// executeARPScan runs the ARP scan and records results.
-func (p *EnumerationPhase) executeARPScan(
-	ctx context.Context,
-	scanCtx context.Context,
-	progress *EnumerationProgress,
-) {
-	progress.SetPhase("arp_scan")
-	if err := p.runARPScan(scanCtx, progress); err != nil {
-		logging.GetLogger().WarnContext(ctx, "ARP scan failed", "error", err)
-		progress.AddError("arp_scan", err)
-	}
-	progress.MarkComplete("arp_scan")
-}
-
-// executeICMPScan runs the ICMP scan and records results.
-func (p *EnumerationPhase) executeICMPScan(
-	ctx context.Context,
-	scanCtx context.Context,
-	progress *EnumerationProgress,
-) {
-	progress.SetPhase("icmp_scan")
-	if err := p.runICMPScan(scanCtx, progress); err != nil {
-		logging.GetLogger().WarnContext(ctx, "ICMP scan failed", "error", err)
-		progress.AddError("icmp_scan", err)
-	}
-	progress.MarkComplete("icmp_scan")
-}
-
-// broadcastDiscoveries sends discovery events for all devices.
-func (p *EnumerationPhase) broadcastDiscoveries(devices []*DiscoveredDevice) {
-	if p.broadcaster == nil {
-		return
-	}
-	for _, device := range devices {
-		p.broadcaster.BroadcastPipelineEvent(PipelineEvent{
-			Type:      EventDeviceDiscovered,
-			Timestamp: time.Now(),
-			Payload: DeviceDiscoveredPayload{
-				IP:      device.IP,
-				MAC:     device.MAC,
-				Vendor:  device.Vendor,
-				Methods: methodsToStrings(device.DiscoveryMethod),
-				IsNew:   true, // TODO: Track new vs existing
-			},
-		})
-	}
-}
-
-// runARPScan performs ARP-based Layer 2 discovery.
-func (p *EnumerationPhase) runARPScan(ctx context.Context, progress *EnumerationProgress) error {
-	// Use the existing device discovery scan which includes ARP
-	if err := p.deviceDiscovery.Scan(ctx); err != nil {
-		return err
-	}
-
-	// Count discovered devices
-	devices := p.deviceDiscovery.GetDevices()
-	for _, d := range devices {
-		if containsMethod(d.DiscoveryMethod, MethodARP) {
-			progress.IncrementDevices()
-		}
-	}
-
-	return nil
-}
-
-// runICMPScan performs ICMP ping sweep on additional subnets.
-func (p *EnumerationPhase) runICMPScan(_ context.Context, _ *EnumerationProgress) error {
-	if len(p.config.AdditionalSubnets) == 0 {
-		return nil
-	}
-
-	// Set additional subnets in the scanner
-	if err := p.deviceDiscovery.SetAdditionalSubnets(p.config.AdditionalSubnets); err != nil {
-		return err
-	}
-
-	// The ARP scan already handles additional subnets with ICMP
-	// This is handled in arp.go's pingSweep method
-	return nil
-}
-
-// EnumerationProgress tracks progress during enumeration.
-// All fields are protected by mu for consistent synchronization.
-type EnumerationProgress struct {
-	mu            sync.RWMutex
-	startTime     time.Time
-	currentPhase  string
-	currentTarget string
-	devicesFound  int
-	errors        []EnumerationError
-	completed     map[string]bool
-}
-
-// EnumerationError records an error during enumeration.
-type EnumerationError struct {
-	Phase string
-	Error error
-	Time  time.Time
-}
-
-// Start initializes progress tracking.
-func (p *EnumerationProgress) Start() {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.startTime = time.Now()
-	p.completed = make(map[string]bool)
-}
-
-// SetPhase updates the current phase.
-func (p *EnumerationProgress) SetPhase(phase string) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.currentPhase = phase
-}
-
-// SetCurrentTarget updates the target being scanned.
-func (p *EnumerationProgress) SetCurrentTarget(target string) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.currentTarget = target
-}
-
-// CurrentTarget returns the current scan target.
-func (p *EnumerationProgress) CurrentTarget() string {
-	p.mu.RLock()
-	defer p.mu.RUnlock()
-	return p.currentTarget
-}
-
-// IncrementDevices adds to the device count.
-func (p *EnumerationProgress) IncrementDevices() {
-	p.mu.Lock()
-	p.devicesFound++
-	p.mu.Unlock()
-}
-
-// DevicesFound returns the number of devices discovered.
-func (p *EnumerationProgress) DevicesFound() int {
-	p.mu.RLock()
-	defer p.mu.RUnlock()
-	return p.devicesFound
-}
-
-// EstimatedTotal returns estimated total devices (for progress calculation).
-func (p *EnumerationProgress) EstimatedTotal() int {
-	// Rough estimate based on typical /24 subnet
-	return enumSubnet24HostCount
-}
-
-// PercentComplete returns completion percentage.
-func (p *EnumerationProgress) PercentComplete() float64 {
-	found := p.DevicesFound()
-	if found == 0 {
-		return 0
-	}
-	// We don't know total, so estimate based on typical discovery rate
-	// After finding devices, we're "mostly done"
-	return float64(found) / float64(p.EstimatedTotal()) * enumPercentComplete
-}
-
-// AddError records an error.
-func (p *EnumerationProgress) AddError(phase string, err error) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.errors = append(p.errors, EnumerationError{
-		Phase: phase,
-		Error: err,
-		Time:  time.Now(),
-	})
-}
-
-// MarkComplete marks a sub-phase as complete.
-func (p *EnumerationProgress) MarkComplete(phase string) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.completed[phase] = true
-}
-
-// Errors returns all recorded errors.
-func (p *EnumerationProgress) Errors() []EnumerationError {
-	p.mu.RLock()
-	defer p.mu.RUnlock()
-	return append([]EnumerationError{}, p.errors...)
-}
-
-// methodsToStrings converts Method slice to string slice.
-func methodsToStrings(methods []Method) []string {
-	result := make([]string, len(methods))
-	for i, m := range methods {
-		result[i] = string(m)
-	}
-	return result
-}
-
-// EnhancedEnumerator provides more thorough enumeration capabilities.
-// It extends the basic enumeration with additional techniques.
-type EnhancedEnumerator struct {
-	deviceDiscovery *DeviceDiscovery
-	pinger          *ICMPPinger
-	config          *EnhancedEnumerationConfig
-}
-
-// EnhancedEnumerationConfig extends EnumerationConfig with additional options.
-type EnhancedEnumerationConfig struct {
-	EnumerationConfig
-
-	// MultiPassARP enables multiple ARP scan passes for reliability.
-	MultiPassARP bool `yaml:"multi_pass_arp" json:"multiPassArp"`
-
-	// ARPPasses is the number of ARP scan passes (if MultiPassARP is true).
-	ARPPasses int `yaml:"arp_passes" json:"arpPasses"`
-
-	// GratuitousARP captures gratuitous ARP announcements.
-	GratuitousARP bool `yaml:"gratuitous_arp" json:"gratuitousArp"`
-
-	// SlowScan uses longer delays between probes (IDS-friendly).
-	SlowScan bool `yaml:"slow_scan" json:"slowScan"`
-
-	// RetryUnresponsive retries devices that didn't respond initially.
-	RetryUnresponsive bool `yaml:"retry_unresponsive" json:"retryUnresponsive"`
-
-	// RetryCount for unresponsive devices.
-	RetryCount int `yaml:"retry_count" json:"retryCount"`
-}
-
-// NewEnhancedEnumerator creates an enhanced enumerator.
-func NewEnhancedEnumerator(
-	deviceDiscovery *DeviceDiscovery,
-	config *EnhancedEnumerationConfig,
-) (*EnhancedEnumerator, error) {
-	pinger, err := NewICMPPinger(config.Timing.PingTimeout)
-	if err != nil {
-		// Non-fatal - ICMP may require elevated privileges
-		logging.GetLogger().Warn("ICMP pinger unavailable", "error", err)
-	}
-
-	return &EnhancedEnumerator{
-		deviceDiscovery: deviceDiscovery,
-		pinger:          pinger,
-		config:          config,
-	}, nil
-}
-
-// EnumerateSubnet performs comprehensive enumeration of a subnet.
-func (e *EnhancedEnumerator) EnumerateSubnet(
-	ctx context.Context,
-	cidr string,
-) ([]*DiscoveredDevice, error) {
-	_, subnet, err := net.ParseCIDR(cidr)
-	if err != nil {
-		return nil, fmt.Errorf("parsing CIDR %s: %w", cidr, err)
-	}
-
-	logging.GetLogger().InfoContext(ctx, "Enhanced enumeration starting", "subnet", cidr)
-
-	hosts := generateHostIPs(subnet)
-	logging.GetLogger().DebugContext(ctx, "Host IPs generated", "count", len(hosts))
-
-	if arpErr := e.performARPScans(ctx); arpErr != nil {
-		return nil, arpErr
-	}
-
-	if e.config.RetryUnresponsive && e.pinger != nil {
-		e.retryUnresponsive(ctx, hosts)
-	}
-
-	return e.deviceDiscovery.GetDevices(), nil
-}
-
-// performARPScans executes either multi-pass or single-pass ARP scanning.
-func (e *EnhancedEnumerator) performARPScans(ctx context.Context) error {
-	if e.config.MultiPassARP && e.config.ARPPasses > 1 {
-		return e.runMultiPassARP(ctx)
-	}
-	return e.deviceDiscovery.Scan(ctx)
-}
-
-// runMultiPassARP performs multiple ARP scan passes for reliability.
-func (e *EnhancedEnumerator) runMultiPassARP(ctx context.Context) error {
-	for pass := range e.config.ARPPasses {
-		if err := e.executeARPPass(ctx, pass); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// executeARPPass runs a single ARP scan pass with optional slow scan delay.
-func (e *EnhancedEnumerator) executeARPPass(ctx context.Context, pass int) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("enumeration cancelled during ARP pass: %w", err)
-	}
-
-	logging.GetLogger().DebugContext(ctx, "ARP pass", "pass", pass+1, "total", e.config.ARPPasses)
-
-	if err := e.applySlowScanDelay(ctx, pass); err != nil {
-		return err
-	}
-
-	if scanErr := e.deviceDiscovery.Scan(ctx); scanErr != nil {
-		logging.GetLogger().WarnContext(ctx, "ARP pass failed", "pass", pass+1, "error", scanErr)
-	}
-	return nil
-}
-
-// applySlowScanDelay waits between passes if slow scan mode is enabled.
-func (e *EnhancedEnumerator) applySlowScanDelay(ctx context.Context, pass int) error {
-	if !e.config.SlowScan || pass == 0 {
-		return nil
-	}
-	select {
-	case <-time.After(enumSlowScanDelayS * time.Second):
-		return nil
-	case <-ctx.Done():
-		return fmt.Errorf("enumeration cancelled during slow scan delay: %w", ctx.Err())
-	}
-}
-
-// retryUnresponsive pings hosts that weren't discovered.
-func (e *EnhancedEnumerator) retryUnresponsive(ctx context.Context, hosts []net.IP) {
-	unresponsive := e.findUnresponsiveHosts(hosts)
-	if len(unresponsive) == 0 {
-		return
-	}
-
-	logging.GetLogger().DebugContext(ctx, "Retrying unresponsive hosts",
-		"count", len(unresponsive),
-		"retries", e.config.RetryCount)
-
-	for range e.config.RetryCount {
-		if ctx.Err() != nil {
-			return
-		}
-
-		unresponsive = e.retryPingSweep(ctx, unresponsive)
-		if len(unresponsive) == 0 {
-			break
-		}
-
-		if !e.waitBetweenRetries(ctx) {
-			return
-		}
-	}
-}
-
-// findUnresponsiveHosts returns hosts that haven't been discovered yet.
-func (e *EnhancedEnumerator) findUnresponsiveHosts(hosts []net.IP) []net.IP {
-	discovered := e.buildDiscoveredSet()
-
-	var unresponsive []net.IP
-	for _, ip := range hosts {
-		if !discovered[ip.String()] {
-			unresponsive = append(unresponsive, ip)
-		}
-	}
-	return unresponsive
-}
-
-// buildDiscoveredSet creates a set of already discovered IP addresses.
-func (e *EnhancedEnumerator) buildDiscoveredSet() map[string]bool {
-	discovered := make(map[string]bool)
-	for _, d := range e.deviceDiscovery.GetDevices() {
-		discovered[d.IP] = true
-	}
-	return discovered
-}
-
-// retryPingSweep performs a ping sweep and returns hosts that are still unresponsive.
-func (e *EnhancedEnumerator) retryPingSweep(ctx context.Context, unresponsive []net.IP) []net.IP {
-	results := e.pinger.PingSweep(ctx, unresponsive, e.config.Timing.PingWorkers)
-	return e.filterStillUnresponsive(results)
-}
-
-// filterStillUnresponsive extracts IPs that remain unreachable from ping results.
-func (e *EnhancedEnumerator) filterStillUnresponsive(results []PingResult) []net.IP {
-	var stillUnresponsive []net.IP
-	for _, r := range results {
-		if !r.Reachable {
-			if ip := net.ParseIP(r.IP); ip != nil {
-				stillUnresponsive = append(stillUnresponsive, ip)
-			}
-		}
-	}
-	return stillUnresponsive
-}
-
-// waitBetweenRetries waits for the retry delay. Returns false if context is cancelled.
-func (e *EnhancedEnumerator) waitBetweenRetries(ctx context.Context) bool {
-	select {
-	case <-time.After(enumRetryDelayMs * time.Millisecond):
-		return true
-	case <-ctx.Done():
-		return false
-	}
-}
-
-// Close releases resources.
-func (e *EnhancedEnumerator) Close() error {
-	if e.pinger != nil {
-		return e.pinger.Close()
-	}
-	return nil
-}
-
-// generateHostIPs returns all host IPs in a subnet (excluding network and broadcast).
-func generateHostIPs(subnet *net.IPNet) []net.IP {
-	ones, bits := subnet.Mask.Size()
-
-	// Prevent integer overflow: limit host bits to avoid shift overflow
-	// For subnets larger than /16 (65534 hosts), cap at 65534
-	hostBits := bits - ones
-	if hostBits <= 0 {
-		return nil // /32 or invalid subnet
-	}
-
-	// Cap at 16 host bits (65534 hosts) to prevent overflow and limit scan size
-	const maxHostBits = 16
-	if hostBits > maxHostBits {
-		hostBits = maxHostBits
-	}
-
-	numHosts := (1 << hostBits) - enumSubnetExcludeNum
-	if numHosts <= 0 {
-		return nil // /31 or smaller
-	}
-
-	baseIP := subnet.IP.Mask(subnet.Mask).To4()
-	if baseIP == nil {
-		return nil
-	}
-
-	hosts := make([]net.IP, 0, numHosts)
-	for i := 1; i <= numHosts; i++ {
-		ip := incrementIP(baseIP, i)
-		if ip != nil {
-			hosts = append(hosts, ip)
-		}
-	}
-
-	return hosts
-}
